requirements.txt
Django
ffmpeg-python
mutagen
yt-dlp
pedalboard
numpy